Update code.py
I've edited the code here to incorporate a periodic initial condition.

My changes begin from line 38 - firstly, I've removed our 2 extra columns.

LINE 40:
Now for our first row, the first cell (grid[0,0]) is kept as zero for now, and the remaining (n-1) cells are randomly generated to be either 1 or zero, as before.

LINE 47:
Next, I used an 'if statement' to ensure that the first cell (grid[0,0]) is made to equal the value of the last cell in the first row (grid[0,1]). So now the first row wraps around on itself. 

[Adjustment that could be made here:
We could adjust this setup to have one extra cell on either side of the first row only, so as to not lose cells on the first and last columns further down - as we do here.]

LINE 77:
Now, our plot contains every cell we work with - no extra columns here to exclude. 

Everything else is left unchanged.